pyproject.toml
[project]
name = "orion_cli"
version = "0.1.0"
description = "CLI for Orion AI agent with LTM and Persona support"
authors = [{ name="Your Name", email="your@email.com" }]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "typer[all]",
    "chromadb",
    "sentence-transformers",
    "pyyaml",
    "rich"
]

[project.scripts]
orion = "orion_cli.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
packages = ["orion_cli"]
